+++ src/consts.ts
@@ -1,0 +1,6 @@
+export enum CommandIds {
+  translateInsertCommand = 'translatorHelper.translateInsert',
+  reverseTranslateInsertCommand = 'translatorHelper.reverseTranslateInsert',
+  translateCommand = 'translatorHelper.translate',
+  copyTranslationTextCommand = 'translatorHelper.copyTranslationText',
+}